Split the import from covidhub into user/groups and data
### Description
Turns out some of the internal CZB projects do not have names that match aspen's expectations.

#### Issue
[ch133318](https://app.clubhouse.io/genepi/story/133318/port-over-internal-samples-from-covidhub-db-into-appropriate-projects-about-500)

### Test plan
imported users and data for marin successfully.
