+++ src/backend/aspen/covidhub_import/__init__.py
@@ -3,4 +3,5 @@ try:
 except ImportError:
     ...
 else:
-    from .implementation import import_project, retrieve_auth0_users  # noqa: F401
+    from .implementation import import_project  # noqa: F401
+    from .import_users import import_project_users, retrieve_auth0_users  # noqa: F401
