+++ src/backend/aspen/covidhub_import/implementation.py
@@ -3,7 +3,6 @@ import json
 import logging
 import re
 import warnings
-from dataclasses import dataclass
 from typing import (
     Any,
     Collection,
@@ -18,13 +17,10 @@ from typing import (
 import boto3
 import pytz
 import tqdm
-from auth0.v3 import authentication as auth0_authentication
-from auth0.v3 import management as auth0_management
 from sqlalchemy import and_, or_
 from sqlalchemy.orm import configure_mappers, joinedload, Session
 
 from aspen.aws.s3 import S3UrlParser
-from aspen.config import config as aspen_config
 from aspen.database.connection import session_scope, SqlAlchemyInterface
 from aspen.database.models import (
     Group,
@@ -34,7 +30,6 @@ from aspen.database.models import (
     RegionType,
     Sample,
     UploadedPathogenGenome,
-    User,
     WorkflowStatusType,
 )
 from aspen.phylo_tree.identifiers import get_names_from_tree
@@ -52,13 +47,6 @@ from covid_database.models.ngs_sample_tracking import (
 )
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class Auth0Entry:
-    nickname: str
-    email: str
-    auth0_token: str
 
 
 def covidhub_interface_from_secret(
@@ -86,28 +74,14 @@ def list_bucket(s3_resource, bucket: str, key_prefix: str) -> Iterator[str]:
         nexttoken = results["NextContinuationToken"]
 
 
-def get_or_make_group(session: Session, name: str, address: str, email: str) -> Group:
-    group = session.query(Group).filter(Group.name == name).one_or_none()
-    if group is None:
-        # copy the project info into the "group"
-        group = Group(
-            name=name,
-            address=address,
-            email=email,
-        )
-        session.add(group)
-
-    return group
-
-
 def import_project(
     interface: SqlAlchemyInterface,
     covidhub_aws_profile: str,
     covidhub_secret_id: str,
     rr_project_id: str,
+    aspen_group_id: int,
     s3_src_prefix: str,
     s3_dst_prefix: str,
-    auth0_usermap: Mapping[str, Auth0Entry],
 ):
     configure_mappers()
     covidhub_interface = covidhub_interface_from_secret(
@@ -126,77 +100,7 @@ def import_project(
             .filter(Project.rr_project_id == rr_project_id)
             .one()
         )
-
-        group, admin_group, cdph_group = (
-            get_or_make_group(
-                session,
-                project.originating_lab,
-                project.originating_address,
-                # FIXME: this needs to be updated.
-                email="fakeemail@dph.gov",
-            ),
-            get_or_make_group(
-                session,
-                "ADMIN",
-                "123 Main St.",
-                # FIXME: this needs to be updated.
-                email="FAKE_EMAIL@gmail.com",
-            ),
-            get_or_make_group(
-                session,
-                "CDPH",
-                "3701 N Freeway Blvd, Sacramento, CA 95834",
-                # FIXME: this needs to be updated.
-                email="cdph.internetadmin@cdph.ca.gov",
-            ),
-        )
-        covidhub_group_to_aspen_group = {
-            project.covidtracker_group.group: group,
-            covidhub_session.query(covidtracker.Group)
-            .filter(covidtracker.Group.name == "Admin")
-            .one(): admin_group,
-            covidhub_session.query(covidtracker.Group)
-            .filter(covidtracker.Group.name == "CDPH")
-            .one(): cdph_group,
-        }
-
-        covidhub_users: Iterable[covidtracker.UsersGroups] = covidhub_session.query(
-            covidtracker.UsersGroups
-        ).filter(
-            or_(
-                covidtracker.UsersGroups.group_id.in_(
-                    covidhub_session.query(covidtracker.Group.id).filter(
-                        covidtracker.Group.name.in_(["Admin", "CDPH"])
-                    )
-                ),
-                covidtracker.UsersGroups.group_id.in_(
-                    covidhub_session.query(covidtracker.GroupToProjects.group_id)
-                    .join(Project)
-                    .filter(Project.rr_project_id == rr_project_id)
-                ),
-            ),
-        )
-
-        # create the users!
-        for covidhub_user in covidhub_users:
-            # try to find the user in the auth0_usermap
-            auth0_user = auth0_usermap.get(covidhub_user.user_id, None)
-            if auth0_user is None:
-                continue
-
-            # try to create this user in the aspen db.
-            user = (
-                session.query(User).filter(User.email == auth0_user.email).one_or_none()
-            )
-            if user is None:
-                user = User()
-
-            user.name = auth0_user.nickname
-            user.email = auth0_user.email
-            user.auth0_user_id = auth0_user.auth0_token
-            user.group_admin = False
-            user.system_admin = covidhub_user.group.name == "Admin"
-            user.group = covidhub_group_to_aspen_group[covidhub_user.group]
+        group: Group = session.query(Group).filter(Group.id == aspen_group_id).one()
 
         logger.info("Retrieving from COVIDHUB DB...")
         consensus_genomes: Collection[ConsensusGenome] = (
@@ -412,37 +316,3 @@ def format_sample_for_dphczbid(
         sample.czb_failed_genome_recovery = True
 
     return sample
-
-
-def retrieve_auth0_users(config: aspen_config.Config) -> Mapping[str, Auth0Entry]:
-    """Retrieves an auth0 token and then retrieve a user list.  Each element of the list
-    is a tuple of (name, email, auth0_token)."""
-    domain = config.AUTH0_DOMAIN
-    client_id = config.AUTH0_MANAGEMENT_CLIENT_ID
-    client_secret = config.AUTH0_MANAGEMENT_CLIENT_SECRET
-
-    get_token_response = auth0_authentication.GetToken(domain)
-    token = get_token_response.client_credentials(
-        client_id, client_secret, "https://{}/api/v2/".format(domain)
-    )
-    mgmt_api_token = token["access_token"]
-    auth0_client = auth0_management.Auth0(domain, mgmt_api_token)
-
-    all_users: MutableMapping[str, Auth0Entry] = dict()
-    page = 0
-    while True:
-        try:
-            users_response = auth0_client.users.list(page=page)
-            if users_response["length"] == 0:
-                # done!
-                break
-
-            # concatenate the results to all_users.
-            for user in users_response["users"]:
-                all_users[user["user_id"]] = Auth0Entry(
-                    user["nickname"], user["email"], user["user_id"]
-                )
-        finally:
-            page += 1
-
-    return all_users
